move NftSaleBottomDrawer to views/global
