+++ packages/atlas/src/views/global/NftSaleBottomDrawer/NftForm/AcceptTerms/index.ts
