+++ packages/atlas/src/views/global/NftSaleBottomDrawer/NftForm/ListingType/ListingType.tsx
