+++ packages/atlas/src/views/global/NftSaleBottomDrawer/NftForm/ListingType/ListingTypes.styles.ts
