+++ packages/atlas/src/views/global/NftSaleBottomDrawer/NftForm/ListingType/index.ts
