+++ packages/atlas/src/views/global/NftSaleBottomDrawer/NftForm/NftForm.styles.ts
