+++ packages/atlas/src/views/global/NftSaleBottomDrawer/NftForm/NftForm.tsx
