+++ packages/atlas/src/views/global/NftSaleBottomDrawer/NftForm/NftForm.types.ts
