+++ packages/atlas/src/views/global/NftSaleBottomDrawer/NftForm/RoyaltiesTooltipFooter.tsx
