+++ packages/atlas/src/views/global/NftSaleBottomDrawer/NftForm/SetUp/SetUp.styles.ts
