+++ packages/atlas/src/views/global/NftSaleBottomDrawer/NftForm/SetUp/index.ts
