+++ packages/atlas/src/views/global/NftSaleBottomDrawer/NftForm/index.ts
