+++ packages/atlas/src/views/global/NftSaleBottomDrawer/NftSaleBottomDrawer.tsx
