+++ packages/atlas/src/views/global/NftSaleBottomDrawer/index.ts
